feat(format): portable FormatDialogContent with explicit colors
Extract the format dialog's title, buttons and body into internal blocks
that share a small state holder, reused by the existing FormatDialog and a
new portable FormatDialogContent that a consumer can host in any container.
Add FormatDialogColors and FormatDialogDefaults.colors() plus a colors
parameter (appended, so existing positional calls keep compiling) on both
entry points; FormatDialogContent applies the title/text content colors
itself and leaves background and elevation to the container. Add the
table-format test source set and Compose UI tests. Also top-align the
Condition tab field list, which was vertically centered.
