+++ table-format/build.gradle.kts
@@ -3,6 +3,7 @@ plugins {
     id("ua.wwind.convention.kmp.target.all")
     id("ua.wwind.convention.compose")
     id("ua.wwind.convention.publishing")
+    id("ua.wwind.convention.test")
 }
 
 kotlin {
@@ -15,6 +16,12 @@ kotlin {
             implementation(libs.colorpicker)
             implementation(libs.reorderable)
             implementation(libs.kotlinx.datetime)
+        }
+        commonTest.dependencies {
+            implementation(libs.compose.ui.test)
+        }
+        jvmTest.dependencies {
+            implementation(compose.desktop.currentOs)
         }
     }
 }
